Création de l'ennemis FireMeteor : une météorite qui génère du feu après explosion. Ajout dans EnemyEntity du créateur (EnemyManager) en attribut. Possibilité dans la méthode takeDamage de Personage de préciser la durée de recovery. Création d'une méthode de création d'ennemi dans enemyManager. Il faudrait l'appeler avec le ThreeStepGenerator. Passage de la gravité en float (j'avais changé les speed en float pour permettre de passer la gravité en float et j'avais oublié de changer l'attribut).

+++ trunk/survivalkid/SurvivalKid/src/com/survivalkid/game/entity/GameEntity.java
@@ -52,7 +52,7 @@ public abstract class GameEntity {
 	protected boolean isJumpingUp;
 	protected boolean isJumpingDown;
 	protected boolean isOnFloor;
-	protected int gravity = 0;
+	protected float gravity = 0;
 	protected int maxSpeedDown = 20;
 
 	/** States. */
@@ -457,7 +457,7 @@ public abstract class GameEntity {
 	/**
 	 * @return the gravity
 	 */
-	public int getGravity() {
+	public float getGravity() {
 		return gravity;
 	}
 
@@ -465,7 +465,7 @@ public abstract class GameEntity {
 	 * @param _isSubjectToGravity
 	 *            the gravity to set
 	 */
-	public void setGravity(int _gravity) {
+	public void setGravity(float _gravity) {
 		this.gravity = _gravity;
 	}
